+++ VCA/app/src/main/res/drawable/button_border2.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+    <selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+
+    <item>
+
+
+    <shape android:shape="rectangle">
+
+    <stroke android:color="#000000" android:width="2dp"/>
+
+</shape>
+
+</item>
+
+</selector>
