+++ VCA/app/src/main/res/layout/activity_audio.xml
@@ -23,7 +23,7 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/clock"
             android:padding="10dp"
             android:text="@string/time" />
@@ -48,7 +48,7 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/weather"
             android:padding="10dp"
             android:text="@string/weather" />
@@ -73,7 +73,7 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/music"
             android:padding="10dp"
             android:text="@string/music" />
