+++ VCA/app/src/main/res/layout/activity_home.xml
@@ -25,7 +25,7 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/lights"
             android:padding="10dp"
             android:text="@string/lights" />
@@ -50,7 +50,7 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/heating"
             android:padding="10dp"
             android:text="@string/heating" />
@@ -75,7 +75,7 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/camera"
             android:padding="10dp"
             android:text="@string/camera" />
