+++ VCA/app/src/main/res/layout/activity_miscellaneous.xml
@@ -25,7 +25,7 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/game"
             android:padding="10dp"
             android:text="@string/game" />
@@ -50,11 +50,33 @@
             android:layout_height="100dp"
             android:layout_gravity="center"
             android:layout_weight="1"
-            android:background="@drawable/button_border"
+            android:background="@drawable/button_border2"
             android:drawableStart="@drawable/calendar"
             android:padding="10dp"
-            android:text="@string/calendar" />
+            android:text="Calendar" >
+
+        </Button>
+
+        <ImageButton
+            android:id="@+id/btnSpeak"
+            android:layout_width="100dp"
+            android:layout_height="100dp"
+            android:layout_margin="0dp"
+            android:layout_marginRight="0dp"
+            android:layout_marginTop="0dp"
+            android:layout_marginLeft="-100dp"
+            android:background="@drawable/button_border2"
+            android:padding="10dp"
+            android:src="@android:drawable/ic_btn_speak_now">
+        </ImageButton>
+
+
+
 
     </LinearLayout>
+
+
+
+
 
 </RelativeLayout>
